+++ autolens/__init__.py
@@ -100,12 +100,12 @@ from .interferometer.model.analysis import AnalysisInterferometer
 from .point.point_dataset import PointDataset
 from .point.fit_point.point_dataset import FitPointDataset
 from .point.fit_point.fluxes import FitFluxes
-from .point.fit_point.positions.abstract import AbstractFitPositionsImagePair
-from .point.fit_point.positions.image_pair import FitPositionsImagePair
-from .point.fit_point.positions.image_pair_all import FitPositionsImagePairAll
-from .point.fit_point.positions.image_pair_repeat import FitPositionsImagePairRepeat
-from .point.fit_point.positions_source import FitPositionsSource
-from .point.fit_point.max_separation import FitPositionsSourceMaxSeparation
+from .point.fit_point.positions.image.abstract import AbstractFitPositionsImagePair
+from .point.fit_point.positions.image.pair import FitPositionsImagePair
+from .point.fit_point.positions.image.pair_all import FitPositionsImagePairAll
+from .point.fit_point.positions.image.pair_repeat import FitPositionsImagePairRepeat
+from .point.fit_point.positions.source.separations import FitPositionsSource
+from .point.fit_point.positions.source.max_separation import FitPositionsSourceMaxSeparation
 from .point.model.analysis import AnalysisPoint
 from .point.solver import PointSolver
 from .quantity.fit_quantity import FitQuantity
