+++ test_autolens/point/fit_point/positions/image/test_pair_repeat.py
