+++ test_autolens/point/fit_point/positions/source/test_separations.py
@@ -12,10 +12,10 @@ def test__two_sets_of_positions__residuals_likelihood_correct():
     noise_map = al.ArrayIrregular([0.5, 1.0])
 
     fit = al.FitPositionsSource(
-        name="point_0", positions=positions, noise_map=noise_map, tracer=tracer
+        name="point_0", data=positions, noise_map=noise_map, tracer=tracer
     )
 
-    assert fit.model_positions.in_list == [(0.0, 1.0), (0.0, 2.0)]
+    assert fit.model_data.in_list == [(0.0, 1.0), (0.0, 2.0)]
     assert fit.noise_map.in_list == [0.5, 1.0]
     assert fit.residual_map.in_list == [1.0, 2.0]
     assert fit.normalized_residual_map.in_list == [1.0 / 0.5, 2.0 / 1.0]
@@ -31,10 +31,10 @@ def test__two_sets_of_positions__residuals_likelihood_correct():
     tracer = al.Tracer(galaxies=[galaxy_mass, galaxy_point_source])
 
     fit = al.FitPositionsSource(
-        name="point_0", positions=positions, noise_map=noise_map, tracer=tracer
+        name="point_0", data=positions, noise_map=noise_map, tracer=tracer
     )
 
-    assert fit.model_positions.in_list == [(0.0, 0.0), (0.0, 1.0)]
+    assert fit.model_data.in_list == [(0.0, 0.0), (0.0, 1.0)]
     assert fit.log_likelihood == pytest.approx(-1.6447298, 1.0e-4)
 
 
@@ -51,16 +51,16 @@ def test__multi_plane_position_solving():
     traced_grids = tracer.traced_grid_2d_list_from(grid=positions)
 
     fit_0 = al.FitPositionsSource(
-        name="point_0", positions=positions, noise_map=noise_map, tracer=tracer
+        name="point_0", data=positions, noise_map=noise_map, tracer=tracer
     )
 
-    assert fit_0.model_positions[0, 1] == pytest.approx(0.326054, 1.0e-1)
-    assert fit_0.model_positions[1, 1] == pytest.approx(1.326054, 1.0e-1)
+    assert fit_0.model_data[0, 1] == pytest.approx(0.326054, 1.0e-1)
+    assert fit_0.model_data[1, 1] == pytest.approx(1.326054, 1.0e-1)
 
-    assert (fit_0.model_positions == traced_grids[1]).all()
+    assert (fit_0.model_data == traced_grids[1]).all()
 
     fit_1 = al.FitPositionsSource(
-        name="point_1", positions=positions, noise_map=noise_map, tracer=tracer
+        name="point_1", data=positions, noise_map=noise_map, tracer=tracer
     )
 
-    assert (fit_1.model_positions == traced_grids[2]).all()
+    assert (fit_1.model_data == traced_grids[2]).all()
